applications_api/serializers.py
from rest_framework import serializers
from applications.models import JobApplication

class JobApplicationSerializer(serializers.ModelSerializer):
    class Meta:
        model = JobApplication
        fields = "__all__"
